organize frontend folders
