+++ frontend/src/components/Header/Login/LoginButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Login} from "./Login"
-import TsmcSushi from "./TsmcSushi"
+import {Login} from "."
+import TsmcSushi from "../TsmcSushi"
 
 function LoginButton() {
   return (
