+++ frontend/src/components/ui/BackgroundImage.tsx
