+++ frontend/src/pages/HomePage/_components/BlueSquare.tsx
