+++ frontend/src/pages/HomePage/_components/Navbar.tsx
