+++ frontend/src/pages/HomePage/_components/SearchBar/Search.png
