+++ frontend/src/pages/HomePage/_components/SearchBar/index.tsx
