+++ frontend/src/pages/HomePage/_components/SearchBar/main.css
