+++ frontend/src/pages/HomePage/_components/Sloganimg.tsx
