+++ frontend/src/pages/HomePage/_components/Specialties.tsx
