+++ frontend/src/pages/MenuFoodPage/MenuFoodPage.tsx
@@ -1,0 +1,12 @@
+import BlueSquareFix from './_components/BlueSquareFix';
+import { Link } from 'react-router-dom';
+
+export default function MenuFoodPage(){
+  return(
+    <>
+      <div>
+        <BlueSquareFix/>
+      </div>
+    </>
+  )
+}
