+++ frontend/src/pages/MenuPage/MenuPage.tsx
@@ -1,13 +1,8 @@
-import BackgroundImage from '../components/BackgroundImage';
-import BlueSquareExtend from '../componentsMenu/BlueSquareExtend';
-
+import BlueSquareExtend from './_components/BlueSquareExtend';
 
 export default function MenuPage(){
   return(
     <>
-      <div className="website">
-        <BackgroundImage />
-      </div>
       <div>
         <div>
           <BlueSquareExtend/>
