+++ frontend/src/pages/MenuPage/_components/MenuFoodType.tsx
