+++ frontend/src/test123.html
@@ -1,4 +1,4 @@
-import * as React from "react";
+<!-- import * as React from "react";
 
 export default function RestaurantComponent() {
   return (
@@ -279,4 +279,4 @@ export default function RestaurantComponent() {
       </footer>
     </main>
   );
-}
+} -->
